refactor: 자동차 이름 객체 생성
