+++ src/main/java/nextstep/domain/Car.java
@@ -1,41 +1,27 @@
 package nextstep.domain;
 
-import java.util.NoSuchElementException;
-import java.util.Objects;
-
 public class Car {
 
-    private static final int MAX_LENGTH = 5;
     private static final int DEFAULT_DISTANCE = 0;
     private static final int MIN_MOVABLE_VALUE = 4;
 
-    private final String name;
+    private final CarName name;
     private int distance;
 
     public Car(String name) {
-        checkName(name);
-        this.name = name;
+        this.name = new CarName(name);
         this.distance = DEFAULT_DISTANCE;
-    }
-
-    private void checkName(String name) {
-        if (Objects.isNull(name) || name.isEmpty()) {
-            throw new NoSuchElementException("[ERROR] 자동차 이름은 빈값이 될 수 없습니다.");
-        }
-        if (name.length() > MAX_LENGTH) {
-            throw new NoSuchElementException("[ERROR] 자동차 이름은 5글자 이하만 가능합니다.");
-        }
     }
 
     public MoveResult move(int number) {
         if (number >= MIN_MOVABLE_VALUE) {
             this.distance++;
         }
-        return new MoveResult(this.name, this.distance);
+        return new MoveResult(name.getName(), this.distance);
     }
 
     public String getName() {
-        return this.name;
+        return name.getName();
     }
 
     public int getMaxDistance(int distance) {
@@ -46,8 +32,8 @@ public class Car {
     }
 
     public StringBuilder appendName(StringBuilder sb, int maxDistance) {
-        if (distance ==maxDistance) {
-            return sb.append(name).append(",");
+        if (distance == maxDistance) {
+            return sb.append(name.getName()).append(",");
         }
         return sb;
     }
