+++ src/test/java/nextstep/domain/CarTest.java
@@ -2,13 +2,9 @@ package nextstep.domain;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import java.util.NoSuchElementException;
-
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 class CarTest {
 
@@ -17,20 +13,6 @@ class CarTest {
     void 자동차_생성(String name) {
         Car car = new Car(name);
         assertThat(car.getName()).isEqualTo(name);
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"javaji"})
-    void 자동차_이름_5자_이내(String name) {
-        assertThatExceptionOfType(NoSuchElementException.class)
-                .isThrownBy(() -> new Car(name));
-    }
-
-    @ParameterizedTest
-    @NullAndEmptySource
-    void 자동차_이름_null(String name) {
-        assertThatExceptionOfType(NoSuchElementException.class)
-                .isThrownBy(() -> new Car(name));
     }
 
     @Test
